Added styles to signin page, removed old styles and signup pages Authenticated the Navigation routes so they wont show when user not logged in Reorder the navigation to look beter for signing out Removed imports we dont need Fixed Style for dashboard to be more consistant with the rest of the app
