+++ client/src/components/Auth/Signin.js
@@ -3,7 +3,10 @@ import firebase from "firebase"
 import { fire } from "./firebaseConfig" // This is being used to provide apiKey to Authentication do not remove
 import StyledFirebaseAuth from "react-firebaseui/StyledFirebaseAuth"
 import { Redirect } from 'react-router-dom'
-export default class Signin extends Component {
+import styles from '../Styles/Signin/styles'
+import {withStyles, CssBaseline, Paper, Typography, Avatar} from '@material-ui/core'
+import LockOutlinedIcon from '@material-ui/icons/LockOutlined';
+class Signin extends Component {
 
 
    uiConfig = {
@@ -21,20 +24,31 @@ export default class Signin extends Component {
 
 
     render(){
+    const { classes } = this.props;
     const { from } = this.props.location.state || { from: { pathname: '/' } }
   
     if (this.props.redirect === true) {
       return <Redirect to={from} />
       
     }
-    console.log(this.props.redirect)
         return(
-            <div>
+          <main className={classes.main}>
+      <CssBaseline />
+      <Paper className={classes.paper}>
+        <Avatar className={classes.avatar}>
+          <LockOutlinedIcon />
+        </Avatar>
+        <Typography component="h1" variant="h5">
+          Sign in
+        </Typography>
               
                 <StyledFirebaseAuth  uiConfig={this.uiConfig}
                                 firebaseAuth={fire}  user={this.props.user}/>
-            </div>
+            </Paper>
+            </main>
     )
     }
         
 }
+
+export default withStyles(styles)(Signin);
