+++ client/src/components/Navigation/index.js
@@ -15,9 +15,7 @@ import {
 } from "@material-ui/core";
 import MenuIcon from "@material-ui/icons/Menu";
 
-//import SigninModal from "./SigninModal";
-//import SignupModal from "./SignupModal";
-// import RegisterLogin from "./RegisterLogin";
+
 import * as ROUTES from "../../constants/routes";
 
 const styles = {
@@ -88,12 +86,27 @@ class Navigation extends React.Component {
               open={open}
               onClose={this.handleMenuClose}
             >
-              <Link component={RouterLink} to={ROUTES.LANDING} style={{textDecoration: "none"}}><MenuItem onClick={this.handleMenuClose}>Home</MenuItem></Link>
-              <Link component={RouterLink} to={ROUTES.DASHBOARD} style={{textDecoration: "none"}}><MenuItem onClick={this.handleMenuClose}>Dashboard</MenuItem></Link>
-              <Link component={RouterLink} to={ROUTES.REPORTS} style={{textDecoration: "none"}}><MenuItem onClick={this.handleMenuClose}>Reports</MenuItem></Link>
-              <Link component={RouterLink} to={ROUTES.SETTINGS} style={{textDecoration: "none"}}><MenuItem onClick={this.handleMenuClose}>Settings</MenuItem></Link>
-              <Link component={RouterLink} to={ROUTES.BILLING} style={{textDecoration: "none"}}><MenuItem onClick={this.handleMenuClose}>Billing</MenuItem></Link>
-              <MenuItem onClick={this.signOut}>Sign Out</MenuItem>
+              <Link component={RouterLink} to={ROUTES.LANDING} style={{textDecoration: "none"}}>
+              <MenuItem onClick={this.handleMenuClose}>Home</MenuItem></Link>
+              <Link component={RouterLink} to={ROUTES.BILLING} style={{textDecoration: "none"}}>
+              <MenuItem onClick={this.handleMenuClose}>Billing</MenuItem></Link>
+              {this.props.authenticated 
+              ?
+              (
+                <div>
+                  <Link component={RouterLink} to={ROUTES.DASHBOARD} style={{textDecoration: "none"}}>
+                  <MenuItem onClick={this.handleMenuClose}>Dashboard</MenuItem></Link>
+                  <Link component={RouterLink} to={ROUTES.REPORTS} style={{textDecoration: "none"}}>
+                  <MenuItem onClick={this.handleMenuClose}>Reports</MenuItem></Link>
+                  <Link component={RouterLink} to={ROUTES.SETTINGS} style={{textDecoration: "none"}}>
+                  <MenuItem onClick={this.handleMenuClose}>Settings</MenuItem></Link>
+                  <MenuItem onClick={this.signOut}>Sign Out</MenuItem>
+
+                </div>
+              )
+              :
+              null
+            }
             </Menu>
         
           </Toolbar>
