--- client/src/components/Styles/Navigation/Signup.js
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-export const SignupModal = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-`;
